requirements.txt
openai==1.69.0
pandas
scikit-learn
fastapi
uvicorn
joblib
pytest
jupyter
python-dotenv
matplotlib
